refactor(fw): Phase 4 SYS module — tear down utils/, R1.2 -Werror (Phase 4 complete)
Final Phase-4 module (7/7). Dismantles the utils/ catch-all per R2.6.

R2.6 — utils/ split by concern:
  - src/utils/utils.c → src/sys/sys_utils.c (timestamp, device-id, CRC32);
    its logging half already went to src/logger/log_text.c (LOG pass).
  - src/utils/{math_utils,timer_callbacks}.c → src/sys/.
  - include/utils/{math_utils,timer_callbacks,types}.h → include/sys/.
  - include/utils/utils.h deleted; its decls split between logger/logger.h
    (logging) and the new sys/sys_utils.h (timing/device/CRC). All include
    sites rewritten accordingly.
  Only src/utils/test_file.* remains (a standalone test/demo, not a module).

R2.1 — include/sys/sys.h umbrella over state, sys_utils, and the folded-in
math_utils / timer_callbacks / types headers.

R1.2 / CONV-06 — sys TUs flipped to full R1.2 + -Werror: heartbeat.c
(void)channel; timer_callbacks.c + sys_utils.c uint64->uint32 casts;
logger.c sign-cast (LOG pass).

Phase 4 is now complete: all 7 owned modules (ACT, EST, SNS, CTRL, COMM,
LOG, SYS) build under the full R1.2 set with -Werror on the arm target.
coding-guidelines R1.2 / R2.1 / R10.3 / CONV-06 flipped to ✅;
warning-rollout.md and requirements §2 finalised.

Verification: arm target links clean (7/7 modules -Werror); host SITL +
full 5-suite ctest green.

+++ src/sys/timer_callbacks.c
@@ -1,7 +1,6 @@
-#include "utils/timer_callbacks.h"
+#include "sys/timer_callbacks.h"
 #include "navhal.h"
 #include "utils.h"
-#include "utils/utils.h"
 #include "variables.h"
 #include <stddef.h>
 #include <stdint.h>
@@ -32,7 +31,7 @@ static void _timer_isr_handler(void) {
 
     if (_interrupt_count >= _callbacks[i].last_call + delay_interrupts) {
       _callbacks[i].callback();
-      _callbacks[i].last_call = _interrupt_count;
+      _callbacks[i].last_call = (uint32_t)_interrupt_count;
     }
   }
 }
